Use __slots__ in Graph and GraphAxis
Enforce the philosophy that all user attributes should be assigned to nodes or edges in the data model.
Graph class specific attributes are static and reserved for use in the class only and thus we can pre-allocate them using __slots__
gaining in performance.
Class copy, getnodes, getedges, __setattr__ and __setitem__ methods are adjusted.